test(calldata): colocate validation fixtures
