+++ suite-common/calldata/src/validation/evm/validateAddress.test.ts
@@ -1,5 +1,5 @@
+import { validateAddressTestCases } from './__fixtures__/mockValidateAddress';
 import { validateAddress } from './address';
-import { validateAddressTestCases } from '../../../mocks/validation/evm/mockValidateAddress';
 
 describe('validateAddress', () => {
     it.each(validateAddressTestCases)('$description', ({ input, context, expected }) => {
