+++ suite-common/calldata/src/validation/shared/__fixtures__/mockBytes32.ts
@@ -1,4 +1,4 @@
-import { type IssueCode } from '../../../src/types/validation';
+import { type IssueCode } from '../../../types/validation';
 
 type TestCase = {
     description: string;
