+++ suite-common/calldata/src/validation/shared/bytes32.test.ts
@@ -1,5 +1,5 @@
+import { findBytes32IssueTestCases } from './__fixtures__/mockBytes32';
 import { findBytes32Issue } from './bytes32';
-import { findBytes32IssueTestCases } from '../../../mocks/validation/shared/mockBytes32';
 
 describe('findBytes32Issue', () => {
     it.each(findBytes32IssueTestCases)('$description', ({ input, expected }) => {
