+++ suite-common/calldata/src/validation/shared/uint16.test.ts
@@ -1,5 +1,5 @@
+import { findUint16OverflowIssueTestCases } from './__fixtures__/mockUint16';
 import { findUint16OverflowIssue } from './uint16';
-import { findUint16OverflowIssueTestCases } from '../../../mocks/validation/shared/mockUint16';
 
 describe('findUint16OverflowIssue', () => {
     it.each(findUint16OverflowIssueTestCases)('$description', ({ input, expected }) => {
