+++ suite-common/calldata/src/validation/shared/uint64.test.ts
@@ -1,5 +1,5 @@
+import { findUint64OverflowIssueTestCases } from './__fixtures__/mockUint64';
 import { findUint64OverflowIssue } from './uint64';
-import { findUint64OverflowIssueTestCases } from '../../../mocks/validation/shared/mockUint64';
 
 describe('findUint64OverflowIssue', () => {
     it.each(findUint64OverflowIssueTestCases)('$description', ({ input, expected }) => {
