+++ suite-common/calldata/src/validation/shared/validateBytes32.test.ts
@@ -1,5 +1,5 @@
+import { validateBytes32TestCases } from './__fixtures__/mockValidateBytes32';
 import { validateBytes32 } from './bytes32';
-import { validateBytes32TestCases } from '../../../mocks/validation/shared/mockValidateBytes32';
 
 describe('validateBytes32', () => {
     it.each(validateBytes32TestCases)('$description', ({ input, expected }) => {
