+++ suite-common/calldata/src/validation/shared/validateUint16.test.ts
@@ -1,5 +1,5 @@
+import { validateUint16TestCases } from './__fixtures__/mockValidateUint16';
 import { validateUint16 } from './uint16';
-import { validateUint16TestCases } from '../../../mocks/validation/shared/mockValidateUint16';
 
 describe('validateUint16', () => {
     it.each(validateUint16TestCases)('$description', ({ input, expected }) => {
