+++ suite-common/calldata/src/validation/tron/validateTronAddress.test.ts
@@ -1,5 +1,5 @@
+import { validateTronAddressTestCases } from './__fixtures__/mockValidateTronAddress';
 import { validateTronAddress } from './address';
-import { validateTronAddressTestCases } from '../../../mocks/validation/tron/mockValidateTronAddress';
 
 describe('validateTronAddress', () => {
     it.each(validateTronAddressTestCases)('$description', ({ input, context, expected }) => {
